Move `messageArgs`
